refactor(platform): harden client for verified x64 operation
- enforce x64 configuration through CMake
- harden pointer and integer conversions for 64-bit safety
- preserve packet, file and ABI-critical layouts
- add x64 architecture and pointer round-trip validation
- harden asset, pack and LZO length handling
- verify Release and Debug x64 builds
- confirm all generated executables are AMD64/PE32+
- confirm zero D3D8/D3D9/D3DX imports
- pass all platform and renderer regression tests
- document the remaining blocked live-runtime validation

+++ src/DumpProto/dump_proto.cpp
@@ -8,6 +8,7 @@
 #include <vector>
 #include <algorithm>
 #include <cstring>
+#include <cstdint>
 #include <io.h>
 #include <direct.h>
 #include <sodium.h>
@@ -163,7 +164,7 @@ enum EItemMisc
 typedef struct SItemLimit
 {
 	BYTE	bType;
-	long	lValue;
+	int32_t	lValue;
 } TItemLimit;
 #pragma pack()
 
@@ -171,7 +172,7 @@ typedef struct SItemLimit
 typedef struct SItemApply
 {
 	BYTE	bType;
-	long	lValue;
+	int32_t	lValue;
 } TItemApply;
 #pragma pack()
 
@@ -198,8 +199,8 @@ typedef struct
 
 	TItemLimit	aLimits[ITEM_LIMIT_MAX_NUM];
 	TItemApply	aApplies[ITEM_APPLY_MAX_NUM];
-	long        alValues[ITEM_VALUES_MAX_NUM];
-	long	alSockets[ITEM_SOCKET_MAX_NUM];
+	int32_t     alValues[ITEM_VALUES_MAX_NUM];
+	int32_t	alSockets[ITEM_SOCKET_MAX_NUM];
 	DWORD	dwRefinedVnum;
 	WORD	wRefineSet;
 	BYTE	bAlterToMagicItemPct;
